requirements.txt
git+https://github.com/openai/CLIP.git
decord==0.6.0
einops==0.4.1
matplotlib==3.5.3
numpy==1.24.2
opencv-python==4.6.0.66
Pillow==9.4.0
tensorboard==2.10.0
tensorboardX==2.5.1
timm==0.4.5
scipy==1.8.1
scikit-learn==1.0
scikit-image==0.19.3
torchmetrics==0.10.2
ftfy
regex
tqdm

